[#844] Moved GsonMapper to util package
