+++ app/src/main/java/org/akvo/flow/domain/apkupdate/ApkUpdateStore.java
@@ -22,6 +22,7 @@ package org.akvo.flow.domain.apkupdate;
 import android.support.annotation.Nullable;
 
 import org.akvo.flow.data.preference.Prefs;
+import org.akvo.flow.util.GsonMapper;
 import org.akvo.flow.util.ConstantUtil;
 import org.akvo.flow.util.PlatformUtil;
