+++ app/src/test/java/org/akvo/flow/domain/apkupdate/ApkUpdateStoreTest.java
@@ -23,6 +23,7 @@ package org.akvo.flow.domain.apkupdate;
 import android.test.suitebuilder.annotation.SmallTest;
 
 import org.akvo.flow.data.preference.Prefs;
+import org.akvo.flow.util.GsonMapper;
 import org.akvo.flow.util.ConstantUtil;
 import org.junit.Before;
 import org.junit.Test;
